Rename profile parent command file
