+++ src/cli/commands/profile/CreateCommand.ts
@@ -6,7 +6,7 @@ import { join } from 'node:path';
 import { Command } from 'commander';
 
 import { isValidProfileId } from '../../../profiles/ProfileLoader.js';
-import type { ProfileCommandDependencies } from './ProfileCommand.js';
+import type { ProfileCommandDependencies } from './Command.js';
 
 export type CreateProfileScope = 'user' | 'project' | 'project-local';
